app/requirements.txt
# requirements.txt - MongoDB version
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
bcrypt==4.1.2  # For password hashing